ENDPOINT Override HACK with Session Update
Endpoint for running on the QPU is currently different that the qvm api.
Upon import of QPUConnection or get_info from pyquil/qpu.py the ENDPOINT
is overwritten with the job/beta endpoint.

A QPUConnection object now updates the header files of the session to
allow access to the QPU.