Changed print functions to work with python2.7
The previous pages of the docs were in python2.7. On this page, there are numerous print functions for both python 2 and 3. If we're using python 2.7, I changed the print functions I found from print() to print. Hope this is helpful.

Example:
on lines 910 and 911 it's  print(X)
but lines 975 and 977 it's print X

Also, there was an empty print function on line 906 after the print function that actually prints something. Was that there on purpose?