Fix various docs typos
- a The QuantumComputer --> The QuantumComputer

The ":ref:`quantum_computer`" expands to "The QuantumComputer", which
means the full phrase expands to "a The QuantumComputer".

The other option is to drop "The" from section heading for the
`quantum_computer` section, but that is in keeping with other adjacent
section headings like "The Quantum Virtual Machine" and "The Quantum
Processing Unit".

- run -> ran

- visa-versa -> vice versa