Fix mathjax macros
Move the mathjax macros out of _static/mathjax_macros.js and into
Sphinx's mathjax_config variable.

Prior to this change, the macros were failing to load for me, which
caused them to be rendered literally ("\ket") in a garish red font.

When I looked in Chrome Developer tools console, it complained:

    Uncaught ReferenceError: MathJax is not defined
        at mathjax_macros.js:1

I think this is due to the fact that Sphinx loads MathJax
asynchronously from a CDN, then loads mathjax_macros.js immediately
after, so it's possible the macros get loaded before MathJax is
defined. Prior to this change, Sphinx was generating HTML like so:

    <script async="async" type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/latest.js?config=TeX-AMS-MML_HTMLorMML"></script>
    <script type="text/x-mathjax-config">MathJax.Hub.Config({"tex2jax": {"inlineMath": [["$", "$"], ["\\(", "\\)"]], "processEscapes": true, "ignoreClass": "document", "processClass": "math|output_area"}})</script>
    <script type="text/javascript" src="_static/mathjax_macros.js"></script>

Additionally, I believe the path specified as the argument to
MathJax.Ajax.loadComplete at the bottom of mathjax_macros.js was
wrong. It specified a path of "[MathJax]/config/local/local.js", but I
think should have been something like "_static/mathjax_macros.js".

After this change, the macros config is injected inline on every page
by Sphinx, alongside the other MathJax config, like so:

    <script async="async" type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/latest.js?config=TeX-AMS-MML_HTMLorMML"></script>
    <script type="text/x-mathjax-config">MathJax.Hub.Config({"TeX": {"Macros": {"sket": ["\\left|\\left. #1 \\right\\rangle\\!\\right\\rangle", 1], "sbra": ["\\left\\langle\\!\\left\\langle #1 \\right.\\right|", 1], "sbraket": ["\\left\\langle\\!\\left\\langle #1 | #2 \\right\\rangle\\!\\right\\rangle", 2], "ket": ["\\left| #1 \\right\\rangle", 1], "bra": ["\\left\\langle #1 \\right|", 1], "braket": ["\\left\\langle #1 | #2 \\right\\rangle", 2], "vec": ["\\text{vec}\\left(#1\\right)", 1], "tr": ["\\text{Tr}\\left(#1\\right)", 1]}}, "tex2jax": {"inlineMath": [["$", "$"], ["\\(", "\\)"]], "processEscapes": true, "ignoreClass": "document", "processClass": "math|output_area"}})</script>

This ensures that the macros get loaded along with the rest of the
config whenever the async MathJax load from the CDN completes.

Sphinx Doc References:

https://www.sphinx-doc.org/en/master/usage/extensions/math.html#confval-mathjax_config

https://docs.mathjax.org/en/latest/configuration.html#using-in-line-configuration-options