Opt-in to protoquil from compiler
This makes, in a backwards-compatible fashion, two changes:

  * A method `QVMCompiler.quil_to_protoquil()` which is almost
    identical to `QVMCompiler.quil_to_native_quil()` except that it
    calls the `quil_to_protoquil` endpoint. It does what it says on
    the tin -- compiles to protoquil, and also populates the metadata
    property of the compiled program.

  * Adds `protoquil` keyword argument to
    `QuantumComputer.compile()`. This enables users to use the above
    protoquil compilation endpoint, and in a backwards-compatible
    fashion.

Prior to this PR the compiler was configured at launch-time to produce
protoquil (via the `-P` option). This could not be changed while the
server was running. That option remains in the compiler, but now there
is the above endpoint that enables the protoquil option as-and-when
the user desires it.