Opt-in to protoquil from compiler
This makes, in a backwards-compatible fashion, two changes:

  * Adds `protoquil` keyword argument to
    `QuantumComputer.compile()`. When this is set to `True` the
    compiler will employ it's protoquil restrictions on input/output.

Prior to this PR the compiler was configured at launch-time to produce
protoquil (via the `-P` option). This could not be changed while the
server was running. That option remains in the compiler, but now there
is the above endpoint that enables the protoquil option as-and-when
the user desires it.

Remove reference to HTTP server

To be removed in an upcoming release.

docstr

appease flake

Note in docs about protoquil compilation

Update changelog for opt-in protoquil

More docs on optional protoquil stuff