The protoquil keywords default to None
This allows us to maintain backwards compatibility. If the server is
started with `-P`, then `protoquil=None` means do what the server
thinks is best. On a QMI this means force protoquil. If the user
manually specifies `protoquil=True` or more importantly
`protoquil=False` then this will override the behaviour of the
server. Since the user has to manually opt-in to this behaviour, we
can be confident that BC is maintained.

Hooray.