Reinstate support for user-defined gates in PyQVM.execute
In #874, PyQVM was changed to remove the restriction for
run-and-measure style programs. In the process, PyQVM.execute lost the
ability to execute programs with user defined gates, which apparently
broke pennylane-forest's PyQVM support.

This commit reinstates the block of code for extracting defined_gates
from the Program passed to PyQVM.execute, and with it the ability to
execute Programs with (non-parametric) defgates.

Fixes #1059

See also: https://github.com/rigetti/pennylane-forest/issues/22